+++ src/model/NewDeal.js
@@ -312,12 +312,6 @@ NewDeal['StatusEnum'] = {
     "lost": "lost",
 
     /**
-     * value: "lost"
-     * @const
-     */
-    "lost": "lost",
-
-    /**
      * value: "deleted"
      * @const
      */