+++ CHANGELOG.md
@@ -8,6 +8,16 @@ For public Changelog covering all changes done to Pipedrive’s API, webhooks an
 
 ## [Unreleased]
 
+## 13.2.5
+## Changed
+- Updated `POST /notes`: parameter `content` is required
+- Updated `POST /organizations`: parameter `name` is required
+- Updated `POST /persons`: parameter `name` is required
+- Updated `DELETE /persons`: parameter `ids` is required
+- Updated `visible_to` parameter for `POST /deals` and `PUT /deals/{id}` to include type `number`
+- Updated `visible_to` parameter for `POST /persons` and `PUT /persons/{id}` to include type `number`
+- Updated `visible_to` parameter for `POST /organizations` and `PUT /organizations/{id}` to include type `number`
+
 ## 13.2.4
 ### Fixed
 - Added quotation marks to fix formatting for Organizations endpoint `visible_to` parameter